Fix lcd state not detected correctly
lcd-toggle does not init state file according to screen state. Hence it
will not work if screen is off and state file does not exist in /tmp.
This commit introduces a mechanism that greps from swaymsg for screen
state and init the state file accordingly.

Also changed to string comparison for $lcd instead of integer
comparison.

Signed-off-by: lamcw <thomas@lamcw.com>